+++ .git-ci/getjuce.sh
@@ -1,0 +1,10 @@
+#!/bin/sh
+
+JUCEFLAVOUR=$1
+OUTDIR=$2
+
+mkdir -p "${OUTDIR}"
+
+curl -o juce.zip https://d30pueezughrda.cloudfront.net/juce/juce-${JUCEFLAVOUR}.zip
+unzip -q juce.zip -d "${OUTDIR}"
+mv "${OUTDIR}"/JUCE/* "${OUTDIR}" || true