+++ AllRADecoder/Source/LoudspeakerTableComponent.h
@@ -32,7 +32,7 @@ class LoudspeakerTableComponent : public Component, public TableListBoxModel
 {
 
 public:
-    LoudspeakerTableComponent(ValueTree& loudspeakers, LoudspeakerVisualizer& visualizer, EnergyDistributionVisualizer& energyVis, UndoManager& undoM) : data(loudspeakers), undoManager(undoM), lspVisualizer(visualizer), engVisualizer(energyVis)
+    LoudspeakerTableComponent(ValueTree& loudspeakers, LoudspeakerVisualizer& visualizer, EnergyDistributionVisualizer& energyVis, UndoManager& undoM, AllRADecoderAudioProcessor& audioProcessor) : data(loudspeakers), undoManager(undoM), processor(audioProcessor), lspVisualizer(visualizer), engVisualizer(energyVis)
     {
         typeFace = getLookAndFeel().getTypefaceForFont(12);
 
@@ -47,7 +47,8 @@ public:
         table.getHeader().addColumn(getAttributeNameForColumnId(4), 4, 45);
         table.getHeader().addColumn(getAttributeNameForColumnId(5), 5, 50);
         table.getHeader().addColumn(getAttributeNameForColumnId(6), 6, 50);
-        table.getHeader().addColumn(getAttributeNameForColumnId(7), 7, 38);
+        table.getHeader().addColumn(getAttributeNameForColumnId(7), 7, 33);
+        table.getHeader().addColumn(getAttributeNameForColumnId(9), 9, 40, 40, 40, TableHeaderComponent::notSortable);
         table.getHeader().addColumn(getAttributeNameForColumnId(8), 8, 60, 60, 60, TableHeaderComponent::notSortable);
 
         table.setHeaderHeight(23);
@@ -58,6 +59,15 @@ public:
 
     ~LoudspeakerTableComponent()
     {
+    }
+
+    void playNoise (const int row)
+    {
+        if (! data.getChild(row).getProperty("Imaginary"))
+        {
+            const int ch = (int) data.getChild(row).getProperty("Channel");
+            processor.playNoiseBurst(ch);
+        }
     }
 
     void selectedRowsChanged (int lastRowSelected) override
@@ -154,6 +164,16 @@ public:
             removeButton->setRowAndColumn (rowNumber, columnId);
             return removeButton;
         }
+        else if (columnId == 9) // Noise
+        {
+            NoiseButton* noiseButton = static_cast<NoiseButton*> (existingComponentToUpdate);
+            if (noiseButton == nullptr)
+                noiseButton = new NoiseButton (*this);
+
+            noiseButton->setRowAndColumn (rowNumber, columnId);
+            noiseButton->setEnabled(! data.getChild(rowNumber).getProperty("Imaginary"));
+            return noiseButton;
+        }
 
 
         // The other columns are editable text columns, for which we use the custom Label component
@@ -178,6 +198,7 @@ public:
             case 6: return "Imaginary";  break;
             case 7: return "Gain"; break;
             case 8: return "Remove"; break;
+            case 9: return "Noise"; break;
             default: return ""; break;
         }
     }
@@ -209,7 +230,10 @@ public:
         else if (data.getChild(rowNumber).getProperty(getAttributeNameForColumnId(columnId)).isDouble())
         {
             const float value = data.getChild(rowNumber).getProperty(getAttributeNameForColumnId(columnId));
-            return String(value, 0);
+            String ret = String(value, 0);
+            if (columnId == 2 || columnId == 3)
+                ret = ret + String(CharPointer_UTF8 (R"(°)"));
+            return ret;
         }
         else return("NaN");
     }
@@ -243,6 +267,7 @@ private:
     ValueTree& data;
     UndoManager& undoManager;
 
+    AllRADecoderAudioProcessor& processor;
     LoudspeakerVisualizer& lspVisualizer;
     EnergyDistributionVisualizer& engVisualizer;
 
@@ -282,7 +307,17 @@ private:
             {
                 const float alpha = isEnabled() ? 1.0f : 0.5f;
 
-                g.setColour (Colours::white);
+                if ((columnId == 4 || columnId == 7) && ! owner.data.getChild(row).getProperty("Imaginary"))
+                    g.setColour (Colours::white.withMultipliedAlpha(0.4f));
+                else
+                {
+                    if (columnId == 5 && owner.data.getChild(row).getProperty("Imaginary"))
+                        g.setColour (Colours::white.withMultipliedAlpha(0.4f));
+                    else
+                        g.setColour (Colours::white);
+                }
+
+
                 g.setFont (getLookAndFeel().getTypefaceForFont(Font(12.0f)));
                 g.setFont (13.f);
 
@@ -346,6 +381,51 @@ private:
         int row, columnId;
     };
 
+    class NoiseButton  : public TextButton
+    {
+    public:
+        NoiseButton (LoudspeakerTableComponent& td)  : owner (td)
+        {
+            setButtonText("Noise");
+            setColour(TextButton::buttonColourId, Colours::green);
+            onClick = [this](){ owner.playNoise(row); };
+        }
+
+        void setRowAndColumn (const int newRow, const int newColumn)
+        {
+            row = newRow;
+            column = newColumn;
+        }
+
+        void paintButton (Graphics& g, bool isMouseOverButton, bool isButtonDown) override
+        {
+            LookAndFeel& lf = getLookAndFeel();
+
+            Rectangle<float> buttonArea(0.0f, 0.0f, getWidth(), getHeight());
+            buttonArea.reduce(2.0f, 2.0f);
+
+            g.setColour(findColour(TextButton::buttonColourId).withMultipliedAlpha(isButtonDown ? 1.0f : isMouseOverButton ? 0.7f : 0.5f));
+            if (isButtonDown)
+                buttonArea.reduce(0.8f, 0.8f);
+            else if (isMouseOverButton)
+                buttonArea.reduce(0.4f, 0.4f);
+
+            g.drawRoundedRectangle(buttonArea, 2.0f, 1.0f);
+
+            buttonArea.reduce(1.5f, 1.5f);
+            g.setColour(findColour(TextButton::buttonColourId).withMultipliedAlpha(isButtonDown ? 1.0f : isMouseOverButton ? 0.5f : 0.2f));
+
+            g.fillRoundedRectangle(buttonArea, 2.0f);
+
+            lf.drawButtonText (g, *this, isMouseOverButton, isButtonDown);
+        }
+
+    private:
+        LoudspeakerTableComponent& owner;
+        int row, column;
+    };
+
+
     class ImaginaryButton  : public Component
     {
     public:
@@ -354,7 +434,10 @@ private:
             addAndMakeVisible(button);
             button.setButtonText("");
             button.setColour(ToggleButton::tickColourId, Colours::orange);
-            button.onClick = [this](){ owner.setBool(columnId, row, button.getToggleState()); };
+            button.onClick = [this](){
+                owner.setBool(columnId, row, button.getToggleState());
+                owner.repaint();
+            };
         }
 
         void mouseDown (const MouseEvent& event) override