+++ SimpleDecoder/Source/PluginEditor.cpp
@@ -30,7 +30,7 @@ SimpleDecoderAudioProcessorEditor::SimpleDecoderAudioProcessorEditor (SimpleDeco
 {
     // ============== BEGIN: essentials ======================
     // set GUI size and lookAndFeel
-    setResizeLimits(670, 300, 1000, 700); // use this to create a resizeable GUI
+    setResizeLimits(670, 300, 1000, 700); // use this to create a resizable GUI
     setLookAndFeel (&globalLaF);
 
     // make title and footer visible, and set the PluginName