+++ ChangeLog.md
@@ -3,6 +3,14 @@ For a list of all changes and bugfixes see the git log.
 
 Please note, that breaking changes are marked with `!!BREAKING CHANGE!!`. They might lead to an unexpected behaviour and might not be compatible with your previous projects without making some adaptions. See the [Breaking changes article](https://plugins.iem.at/docs/breakingchanges/) for more information.
 
+## v1.3.1
+- plug-in specific changes
+    - **Energy**Visualizer
+        - using a perceptually motivated colormap
+    - **AllRA**Decoder
+        - clicking 'add loudspeaker' while holding the alt-key adds an imaginary loudspeaker at the nadir
+        - added noise generator for testing correct signal routing
+
 ## v1.3.0
 - general changes
     - unity gain normalization has moved fully to the decoding stage:  (`!!BREAKING CHANGE!!`)
@@ -20,7 +28,7 @@ Please note, that breaking changes are marked with `!!BREAKING CHANGE!!`. They m
         - performance improvements
         - new binaural filters
     - **Dual**Delay
-        - fixed high CPU load which occured occasionally
+        - fixed high CPU load which occurred occasionally
     - **Matrix**Multiplier
         - smaller GUI changes
     - **Multi**Encoder